feat(titus): add notifications, move env resolver
