+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/resource/SecurityGroupHandler.kt
@@ -48,6 +48,7 @@ import com.netflix.spinnaker.keel.model.Moniker
 import com.netflix.spinnaker.keel.model.OrchestrationRequest
 import com.netflix.spinnaker.keel.model.OrchestrationTrigger
 import com.netflix.spinnaker.keel.orca.OrcaService
+import com.netflix.spinnaker.keel.plugin.EnvironmentResolver
 import com.netflix.spinnaker.keel.plugin.Resolver
 import com.netflix.spinnaker.keel.plugin.ResourceHandler
 import com.netflix.spinnaker.keel.retrofit.isNotFound
@@ -285,7 +286,8 @@ class SecurityGroupHandler(
       moniker = Moniker(app = moniker.app, stack = moniker.stack, detail = moniker.detail),
       location = SecurityGroup.Location(
         account = accountName,
-        vpc = vpcId?.let { cloudDriverCache.networkBy(it).name } ?: error("Only security groups in a VPC are supported"),
+        vpc = vpcId?.let { cloudDriverCache.networkBy(it).name }
+          ?: error("Only security groups in a VPC are supported"),
         region = region
       ),
       description = description,
