+++ keel-plugin/keel-plugin.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
   api(project(":keel-core"))
+  implementation(project(":keel-orca"))
 
   testImplementation("dev.minutest:minutest")
   testImplementation("io.strikt:strikt-core")
