+++ keel-titus-plugin/src/main/kotlin/com/netflix/spinnaker/config/TitusConfig.kt
@@ -22,6 +22,7 @@ import com.netflix.spinnaker.keel.api.titus.cluster.TitusClusterHandler
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverCache
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverService
 import com.netflix.spinnaker.keel.orca.OrcaService
+import com.netflix.spinnaker.keel.plugin.EnvironmentResolver
 import com.netflix.spinnaker.keel.plugin.Resolver
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.ApplicationEventPublisher
@@ -38,6 +39,7 @@ class TitusConfig {
     cloudDriverCache: CloudDriverCache,
     orcaService: OrcaService,
     clock: Clock,
+    environmentResolver: EnvironmentResolver,
     publisher: ApplicationEventPublisher,
     objectMapper: ObjectMapper,
     resolvers: List<Resolver<*>>
@@ -46,6 +48,7 @@ class TitusConfig {
     cloudDriverCache,
     orcaService,
     clock,
+    environmentResolver,
     publisher,
     objectMapper,
     resolvers
